Pass test filters to python's unittest, update the skipCurrentTest logic, and handle malformed regex
Pass filters to python's unittest by adding the "-k" flag along with the filter to the command.
If class#method filters are used, parse out only the test name and pass that as a filter.
The parser is still useful to filter tests after python's unittest has filtered them, to maintain the same behavior for Class#method syntax.
Update the python test results parser to not skip tests when the fnMatch pattern is used.
If a filter doesn't match either the class#method syntax or fnMatch pattern syntax, match it broadly as a pattern.
Add try/catch blocks for malformed regex filters.

Bug: 338116157
Test: m tradefed tradefed-all && ./javatests/run_tradefed_tests.sh --class com.android.tradefed.testtype.PythonUnitTestResultParserTest
Test: ./javatests/run_tradefed_tests.sh --class com.android.tradefed.testtype.python.PythonBinaryHostTestTest
Test: atest-dev adb_integration_test_device --test-filter SocketTest

Change-Id: I618b112caf17545143132a5fe8ff50d8161745fa