trusty: app_mgr: delay app restart
Previously app restart delays happened after app thread creation, and in
the application thread.  This means the app resources are held during
the delay, and have to be allocated immediately after app exit.

It is possible that shared resources - specifically memory references -  may not be cleared up until some other task are able to react to a task exit. In such a case it is perferrable to delay TA creation and restart until other TAs have had a chance to execute.  Therefore pull the restart delay from the TA thread and into the app_mgr itself.

Bug: 353306393
Test: build.py qemu-generic-arm64-test-debug
Test: on device
Change-Id: I3dea7775c7260079ddd44cc667a08b0480e8ff03