update_engine: change notication order
in the case that update_engine has the state of
ErrorCode::kUpdatedButNotActive, update_engine currently first reports
an IDLE state, before reporting reporting the error. This leads to a
period in between the two reports ambiguous as to what the state of the
OTA is. Moving the report of errorcode ahead to clarify this.

Test: th
Bug: 348514779
Change-Id: I957332cfa576f2eca2fe899e3e5828b389868820