Add note to failure in vts_treble_no_hidl about FCM level
We base the registration of HIDL framework/system services off of the
Framework Compatibiltiy Matrix level that a device is targetting. So in
order for the device to have no HIDL, it must be targetting FCM 202404
or later.
That is accomplished by setting `target-level="202404"` or later in the
device manifest file.

Test: none
Bug: 352265405
Change-Id: Ib678087e7202e17bd08a43e4ea35893cba1c5b8f