Fix 32-bit tests for `arm_v7_v8`.
Make gtest/run-test target platform dependencies explicit.
The `toybox` target listed in the `buildbot-build.sh` has
been indirectly including the required `libprocessgroup.so`
and `libselinux.so` for the primary architecture but we were
missing them for the secondary architecture and therefore
failing to compile the corresponding test boot image. As the
`ART_TARGET_PLATFORM_DEPENDENCIES` variable used for Golem
builds was the only place mentioning the `libprocessgroup`,
we add it to the gtest/run-test depenencies and add support
for the secondary architecture to this variable. The ART
module already ships with an internal `liblz4` library, so
we remove that unnecessary dependency.

Note that we currently do not build the following gtests
for the secondary architecture:
    art_artd_tests
    art_libartservice_tests
    art_libarttools_tests
    art_odrefresh_tests
Other gtests and all run-tests run on both architectures.

Test: run-gtests.sh  # arm_v7_v8
Test: testrunner.py --target --optimizing  # arm_v7_v8
Change-Id: I6bee1f1f777f9e114035104df93e19370ec90fbc